fix up physics inconsistencies with speed multiplier

+++ Character.h
@@ -81,8 +81,8 @@ protected:
 
     CharacterType type;
     CharacterState state;
-    int x; // X coordinate
-    int y; // Y coordinate
+    float x; // X coordinate
+    float y; // Y coordinate
     
     int lives;
 public:
